feat: Switch build system (esbuild)
closes #20
closes #14
closes #13
