--- .parcelrc
@@ -1,3 +1,0 @@
-{
-  "extends": "@parcel/config-webextension"
-}
