--- .sassrc
@@ -1,3 +1,0 @@
-{
-    "includePaths": ["node_modules"]
-}
