+++ package.json
@@ -2,15 +2,14 @@
   "name": "aws-role-switch",
   "version": "0.2.0",
   "license": "MIT",
-  "browserslist": "> 0.5%, last 20 versions, not dead",
   "scripts": {
     "test": "jest --colors",
     "lint": "eslint . --cache",
     "type-check": "tsc --noEmit",
     "start": "run-p watch ext:run",
     "clean": "rimraf dist web-ext-artifacts",
-    "watch": "parcel watch src/manifest.json",
-    "build": "parcel build src/manifest.json --no-source-maps --no-cache",
+    "watch": "./scripts/build.ts --watch",
+    "build": "./scripts/build.ts --minify",
     "bundle": "run-s clean build ext:build",
     "ext:run": "web-ext run -s dist",
     "ext:lint": "web-ext lint -s dist",
@@ -22,6 +21,7 @@
     "@codemirror/legacy-modes": "^6.1.0",
     "@codemirror/view": "^6.0.2",
     "@uiw/react-codemirror": "^4.10.4",
+    "buffer": "^6.0.3",
     "colortranslator": "^1.9.2",
     "js-ini": "^1.5.1",
     "lodash": "^4.17.21",
@@ -33,15 +33,9 @@
     "webextension-polyfill": "^0.9.0",
     "zod": "^3.17.3"
   },
-  "@parcel/bundler-default": {
-    "minBundles": 10000,
-    "minBundleSize": 3000,
-    "maxParallelRequests": 20
-  },
   "devDependencies": {
     "@faker-js/faker": "^7.3.0",
-    "@parcel/config-webextension": "^2.6.2",
-    "@parcel/transformer-sass": "^2.6.2",
+    "@types/chrome": "^0.0.193",
     "@types/jest": "^28.1.4",
     "@types/lodash": "^4.14.182",
     "@types/marked": "^4.0.3",
@@ -51,7 +45,6 @@
     "@types/webextension-polyfill": "^0.9.0",
     "@typescript-eslint/eslint-plugin": "^5.30.5",
     "@typescript-eslint/parser": "^5.30.5",
-    "buffer": "^6.0.3",
     "esbuild": "^0.14.49",
     "esbuild-plugin-copy": "^1.3.0",
     "esbuild-sass-plugin": "^2.2.6",
@@ -62,8 +55,6 @@
     "jest": "^28.1.2",
     "jest-environment-jsdom": "^28.1.2",
     "npm-run-all": "^4.1.5",
-    "parcel": "^2.6.2",
-    "process": "^0.11.10",
     "rimraf": "^3.0.2",
     "ts-jest": "^28.0.5",
     "ts-node": "^10.8.2",
