+++ scripts/build.ts
@@ -1,37 +1,42 @@
 #!/usr/bin/env -S TS_NODE_TRANSPILE_ONLY=true npx ts-node
 
-import { build } from 'esbuild';
+import {
+  build,
+  BuildOptions,
+} from 'esbuild';
 import { copy } from 'esbuild-plugin-copy';
 import { sassPlugin } from 'esbuild-sass-plugin';
 
-const watch = false;
-const minify = false;
+import { mergeJsons } from '../src/common/build/esbuild-plugins';
 
-// popup and options page
-build({
-  entryPoints: [
-    'src/popup/popup.tsx',
-    'src/popup/popup.scss',
-    // 'src/options/options.tsx',
-    // 'src/options/options.scss',
-  ],
+const defaultOps: BuildOptions = {
+  watch: !!process.argv.includes('--watch'),
+  minify: !!process.argv.includes('--minify'),
+  bundle: true,
+  outdir: 'dist',
+  target: 'ES2018',
+  format: 'cjs',
+  sourcemap: 'external',
   loader: {
     '.svg': 'file',
     '.md': 'file',
   },
-  watch,
-  minify,
-  bundle: true,
-  outdir: 'dist',
-  format: 'cjs',
-  sourcemap: 'external',
+};
+
+// popup and options page
+build({
+  ...defaultOps,
+  entryPoints: [
+    'src/popup/popup.tsx',
+    'src/popup/popup.scss',
+    'src/options/options.tsx',
+    'src/options/options.scss',
+  ],
   plugins: [
+    mergeJsons(['src/manifest.json', 'src/manifest.android.json']),
     sassPlugin(),
     copy({
       assets: [{
-        from: ['./src/manifest.json'],
-        to: ['.'],
-      }, {
         from: ['src/assets/*'],
         to: ['./assets/*']
       }, {
@@ -45,17 +50,11 @@ build({
 
 // content/background scripts
 build({
+  ...defaultOps,
+  minify: false,
   entryPoints: [
     'src/content/aws-console.ts',
     'src/content/aws-signin.ts',
     'src/background/background.ts',
   ],
-  watch,
-  bundle: true,
-  outdir: 'dist',
-  target: 'ES2018',
-  format: 'cjs',
-  sourcemap: 'external',
-  // do not minify background/content scripts for readability
-  minify: false,
 });
