+++ src/common/config/gzip.ts
@@ -1,7 +1,11 @@
 import Pako from 'pako';
 
+import { Buffer } from 'buffer';
+
 export const compressConfig = (configFile: string): string => {
   const compressed = Pako.deflate(configFile);
+  // TODO: get rid of Buffer with btoa
+  // https://stackoverflow.com/questions/12710001/how-to-convert-uint8-array-to-base64-encoded-string
   return Buffer.from(compressed).toString('base64');
 };
