+++ src/manifest.android.json
@@ -1,0 +1,14 @@
+{
+  "manifest_version": 2,
+  "background": {
+    "scripts": [
+      "background/background.js"
+    ]
+  },
+  "browser_specific_settings": {
+    "gecko": {
+      "id": "{31f7b254-7ac9-4f3a-ae3c-ef67ea153e4a}",
+      "strict_min_version": "57.0"
+    }
+  }
+}
